Moved RemotePostCategory extension to separate file
