+++ WordPress/WordPressShareExtension/AppExtensionsService.swift
@@ -1,21 +1,6 @@
 import CoreData
 import WordPressKit
 
-extension RemotePostCategory {
-    static func remotePostCategoriesFromString(_ categories: String?) -> [RemotePostCategory]? {
-        guard let categories = categories, !categories.isEmpty else {
-            return nil
-        }
-
-        let remotePostcategories: [RemotePostCategory] = categories.arrayOfTags().flatMap({Int($0)}).map({
-            let remoteCat = RemotePostCategory()
-            remoteCat.categoryID = NSNumber(value: $0)
-            return remoteCat
-        })
-        return remotePostcategories
-    }
-}
-
 /// Provides site fetching and post/media uploading functionality to app extensions.
 ///
 class AppExtensionsService {
