+++ WordPress/WordPressShareExtension/RemotePostCategory+Extensions.swift
@@ -1,0 +1,17 @@
+import Foundation
+import WordPressKit
+
+extension RemotePostCategory {
+    static func remotePostCategoriesFromString(_ categories: String?) -> [RemotePostCategory]? {
+        guard let categories = categories, !categories.isEmpty else {
+            return nil
+        }
+
+        let remotePostcategories: [RemotePostCategory] = categories.arrayOfTags().flatMap({Int($0)}).map({
+            let remoteCat = RemotePostCategory()
+            remoteCat.categoryID = NSNumber(value: $0)
+            return remoteCat
+        })
+        return remotePostcategories
+    }
+}
